api/requirements.txt
gunicorn
flask
mysql-connector-python
flask-cors